Made a mapping of relationships to types
