--- src/TestingSystem.Common/TestingSystem.BL/BusinessObjects/AnswerBusinessObject.cs
@@ -1,10 +1,0 @@
-﻿namespace TestingSystem.BL.BusinessObjects
-{
-    public sealed class AnswerBusinessObject
-    {
-        public int AnswerId { get; set; }
-        public int QuestionId { get; set; }
-        public string Text { get; set; }
-        public bool IsRemoved { get; set; }
-    }
-}
