--- src/TestingSystem.Common/TestingSystem.BL/BusinessObjects/GroupBusinessObject.cs
@@ -1,9 +1,0 @@
-﻿namespace TestingSystem.BL.BusinessObjects
-{
-    public sealed class GroupBusinessObject
-    {
-        public int GroupId { get; set; }
-        public string Name { get; set; }
-        public bool IsRemoved { get; set; }
-    }
-}
