--- src/TestingSystem.Common/TestingSystem.BL/BusinessObjects/RoleBusinessObject.cs
@@ -1,9 +1,0 @@
-﻿namespace TestingSystem.BL.BusinessObjects
-{
-    public sealed class RoleBusinessObject
-    {
-        public int RoleId { get; set; }
-        public string Name { get; set; }
-        public bool IsRemoved { get; set; }
-    }
-}
