--- src/TestingSystem.Common/TestingSystem.BL/BusinessObjects/TestBusinessObject.cs
@@ -1,10 +1,0 @@
-﻿namespace TestingSystem.BL.BusinessObjects
-{
-    public sealed class TestBusinessObject
-    {
-        public int TestId { get; set; }
-        public string Title { get; set; }
-        public string Description { get; set; }
-        public bool IsRemoved { get; set; }
-    }
-}
