--- src/TestingSystem.Common/TestingSystem.BL/BusinessObjects/UserBusinessObject.cs
@@ -1,21 +1,0 @@
-﻿using System;
-
-namespace TestingSystem.BL.BusinessObjects
-{
-    public sealed class UserBusinessObject
-    {
-        public int UserId { get; set; }
-        public int? GroupId { get; set; }
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public string MiddleName { get; set; }
-        public DateTime Birthday { get; set; }
-        public string Email { get; set; }
-        public bool IsEmailVerified { get; set; }
-        public string PhoneNumber { get; set; }
-        public bool IsPhoneVerified { get; set; }
-        public string Login { get; set; }
-        public string Password { get; set; }
-        public bool IsRemoved { get; set; }
-    }
-}
