--- src/TestingSystem.Common/TestingSystem.BL/BusinessObjects/UserRolesBusinessObject.cs
@@ -1,9 +1,0 @@
-﻿namespace TestingSystem.BL.BusinessObjects
-{
-    public sealed class UserRolesBusinessObject
-    {
-        public int UserId { get; set; }
-        public int RoleId { get; set; }
-        public bool IsRemoved { get; set; }
-    }
-}
