--- src/TestingSystem.Common/TestingSystem.BL/BusinessServices/BaseBusinessService.cs
@@ -1,15 +1,0 @@
-﻿using TestingSystem.DAL.DataObjects.Context;
-using TestingSystem.DAL.DataServices;
-
-namespace TestingSystem.BL.BusinessServices
-{
-    public class BaseBusinessService
-    {
-        protected readonly DatabaseDataService databaseDataService;
-
-        public BaseBusinessService()
-        {
-            databaseDataService = new DatabaseDataService(new BaseContext("EntitiesContext"));
-        }
-    }
-}
