--- src/TestingSystem.Common/TestingSystem.BL/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="AutoMapper" version="9.0.0" targetFramework="net471" />
-  <package id="EntityFramework" version="6.3.0" targetFramework="net471" />
-</packages>
