+++ src/TestingSystem.Common/TestingSystem.Common.DAL/DataObjects/Contexts/EntitiesContext.cs
@@ -1,7 +1,10 @@
-namespace TestingSystem.DAL.DataObjects
+namespace TestingSystem.Common.DAL.DataObjects
 {
+    using System;
     using System.Data.Entity;
-    using TestingSystem.DAL.DataObjects.Context;
+    using System.ComponentModel.DataAnnotations.Schema;
+    using System.Linq;
+    using TestingSystem.Common.DAL.DataObjects.Contexts;
 
     public partial class EntitiesContext : BaseContext
     {
