+++ src/TestingSystem.Common/TestingSystem.Common.DAL/DataObjects/Group.cs
@@ -1,26 +1,25 @@
-namespace TestingSystem.DAL.DataObjects
+namespace TestingSystem.Common.DAL.DataObjects
 {
+    using System;
     using System.Collections.Generic;
     using System.ComponentModel.DataAnnotations;
+    using System.ComponentModel.DataAnnotations.Schema;
+    using System.Data.Entity.Spatial;
 
-    public partial class Test
+    public partial class Group
     {
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2214:DoNotCallOverridableMethodsInConstructors")]
-        public Test()
+        public Group()
         {
             GroupTests = new HashSet<GroupTest>();
-            Questions = new HashSet<Question>();
-            StudentTests = new HashSet<StudentTest>();
+            Users = new HashSet<User>();
         }
 
-        public int TestId { get; set; }
+        public int GroupId { get; set; }
 
         [Required]
-        [StringLength(64)]
-        public string Title { get; set; }
-
-        [StringLength(4000)]
-        public string Description { get; set; }
+        [StringLength(13)]
+        public string Name { get; set; }
 
         public bool IsRemoved { get; set; }
 
@@ -28,9 +27,6 @@ namespace TestingSystem.DAL.DataObjects
         public virtual ICollection<GroupTest> GroupTests { get; set; }
 
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
-        public virtual ICollection<Question> Questions { get; set; }
-
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
-        public virtual ICollection<StudentTest> StudentTests { get; set; }
+        public virtual ICollection<User> Users { get; set; }
     }
 }
