+++ src/TestingSystem.Common/TestingSystem.Common.DAL/DataObjects/GroupTest.cs
@@ -1,7 +1,10 @@
-namespace TestingSystem.DAL.DataObjects
+namespace TestingSystem.Common.DAL.DataObjects
 {
     using System;
+    using System.Collections.Generic;
     using System.ComponentModel.DataAnnotations;
+    using System.ComponentModel.DataAnnotations.Schema;
+    using System.Data.Entity.Spatial;
 
     public partial class GroupTest
     {
