+++ src/TestingSystem.Common/TestingSystem.Common.DAL/Properties/AssemblyInfo.cs
@@ -1,14 +1,15 @@
 ﻿using System.Reflection;
+using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 // General Information about an assembly is controlled through the following
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("TestingSystem.BL")]
+[assembly: AssemblyTitle("TestingSystem.Common.DAL")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("TestingSystem.BL")]
+[assembly: AssemblyProduct("TestingSystem.Common.DAL")]
 [assembly: AssemblyCopyright("Copyright ©  2019")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
@@ -19,7 +20,7 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("ca335478-3f1a-419d-bce1-50643851a157")]
+[assembly: Guid("f82bbdc7-b2d1-4b29-ad47-1d28c24b210f")]
 
 // Version information for an assembly consists of the following four values:
 //
