+++ src/TestingSystem.Common/TestingSystem.Common.DAL/TestingSystem.Common.DAL.csproj
@@ -5,11 +5,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{0A61B58C-F0B3-4A1C-922F-EE865FC9770E}</ProjectGuid>
+    <ProjectGuid>{F82BBDC7-B2D1-4B29-AD47-1D28C24B210F}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>TestingSystem.DAL</RootNamespace>
-    <AssemblyName>TestingSystem.DAL</AssemblyName>
+    <RootNamespace>TestingSystem.Common.DAL</RootNamespace>
+    <AssemblyName>TestingSystem.Common.DAL</AssemblyName>
     <TargetFrameworkVersion>v4.7.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <Deterministic>true</Deterministic>
@@ -43,8 +43,6 @@
     <Reference Include="System" />
     <Reference Include="System.ComponentModel.DataAnnotations" />
     <Reference Include="System.Core" />
-    <Reference Include="System.Runtime.Serialization" />
-    <Reference Include="System.Security" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
     <Reference Include="Microsoft.CSharp" />
@@ -53,28 +51,15 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="DataObjects\Answer.cs" />
-    <Compile Include="DataObjects\Context\BaseContext.cs" />
-    <Compile Include="DataObjects\Context\EntitiesContext.cs" />
-    <Compile Include="DataObjects\Group.cs" />
-    <Compile Include="DataObjects\GroupTest.cs" />
-    <Compile Include="DataObjects\Question.cs" />
-    <Compile Include="DataObjects\Role.cs" />
-    <Compile Include="DataObjects\StudentTest.cs" />
-    <Compile Include="DataObjects\Test.cs" />
-    <Compile Include="DataObjects\User.cs" />
-    <Compile Include="DataObjects\UserRole.cs" />
-    <Compile Include="DataServices\Database\Infrastructure\BaseDatabaseDataService.cs" />
-    <Compile Include="DataServices\Database\DatabaseDataService.cs" />
-    <Compile Include="DataServices\Database\Infrastructure\IDatabaseDataService.cs" />
+    <Compile Include="DataObjects\Contexts\BaseContext.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="DataServices\" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
     <None Include="packages.config" />
-  </ItemGroup>
-  <ItemGroup>
-    <Service Include="{508349B6-6B84-4DF5-91F0-309BEEBAD82D}" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
