+++ src/TestingSystem.Common/TestingSystem.Common.sln
@@ -3,9 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29418.71
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestingSystem.DAL", "TestingSystem.DAL\TestingSystem.DAL.csproj", "{0A61B58C-F0B3-4A1C-922F-EE865FC9770E}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestingSystem.BL", "TestingSystem.BL\TestingSystem.BL.csproj", "{CA335478-3F1A-419D-BCE1-50643851A157}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestingSystem.Common.DAL", "TestingSystem.Common.DAL\TestingSystem.Common.DAL.csproj", "{F82BBDC7-B2D1-4B29-AD47-1D28C24B210F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,14 +11,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{0A61B58C-F0B3-4A1C-922F-EE865FC9770E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0A61B58C-F0B3-4A1C-922F-EE865FC9770E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0A61B58C-F0B3-4A1C-922F-EE865FC9770E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0A61B58C-F0B3-4A1C-922F-EE865FC9770E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CA335478-3F1A-419D-BCE1-50643851A157}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CA335478-3F1A-419D-BCE1-50643851A157}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CA335478-3F1A-419D-BCE1-50643851A157}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CA335478-3F1A-419D-BCE1-50643851A157}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F82BBDC7-B2D1-4B29-AD47-1D28C24B210F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F82BBDC7-B2D1-4B29-AD47-1D28C24B210F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F82BBDC7-B2D1-4B29-AD47-1D28C24B210F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F82BBDC7-B2D1-4B29-AD47-1D28C24B210F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
